docker: use multi-stage build
By leveraging docker's multi-stage builds, the size of the final mq2prom
docker image is reduced from 409MB to 21.2MB.

This also adds extra hardening to the image, because a lot less packages
and intermediate artifacts are included as part of it.

As part of this commit, the go version used to build the binary inside
the docker image has been bumped to 1.20.11.

Fixes: #11

Signed-off-by: Juan Vidal Allende <juan.vidal1@ibm.com>